refactor(phase-19)(pr-4a): reading 디렉터리 승격
progression/reading.go (652줄) → progression/reading/ 디렉터리 분할.

파일 구조:
- module.go (123줄) — 구조체, NewReadingModule, Init, Cleanup, engine.Module assertion, init()
- config.go (44줄) — readingConfig, readingLineConfig, Schema
- state.go (104줄) — ReadingState, ReadingStateWire, GetState, GetReadingStateWire, BuildState, BuildStateFor
- handlers.go (352줄) — HandleMessage, HandleAdvance, HandleVoiceEnded, HandlePlayerLeft, HandlePlayerRejoined
- helpers.go (53줄) — readingStatus, validAdvanceBy, resolveAdvanceBy, resolveVoiceID
- phase.go (20줄) — OnPhaseEnter, OnPhaseExit
- reading_test.go — 기존 테스트 이관 (package rename)

외부 caller 경로 갱신:
- bridge/reading_bridge.go, reading_bridge_test.go → progression/reading import
- ws/reading_handler.go — 주석만 갱신 (직접 import 없음)

progression/register.go 신설: reading 하위 package blank-import으로 init() 보장.

Factory 서명, engine.Register 방식, PlayerAware/PublicStateMarker 모두 불변.
F-go-3·4 해소 (500+ 파일 9→8).
