+++ apps/server/internal/module/progression/reading/handlers.go
@@ -1,4 +1,4 @@
-package progression
+package reading
 
 import (
 	"context"
@@ -6,100 +6,11 @@ import (
 	"fmt"
 	"net/http"
 	"strings"
-	"sync"
 
 	"github.com/google/uuid"
 	"github.com/mmp-platform/server/internal/apperror"
 	"github.com/mmp-platform/server/internal/engine"
 )
-
-// ReadingModule manages script reading progression with line-by-line advancement.
-type ReadingModule struct {
-	mu   sync.RWMutex
-	deps engine.ModuleDeps
-
-	// config
-	advanceMode    string // "gm", "auto", "player"
-	defaultVoiceID string
-	bgmId          string
-	lines          []readingLineConfig
-
-	// state
-	currentLineIndex int
-	totalLines       int
-	isActive         bool
-	status           readingStatus
-	pausedReason     string
-}
-
-// readingStatus describes whether the reading flow is currently advancing,
-// halted (waiting on a missing player), or not yet/no longer active.
-type readingStatus string
-
-const (
-	readingStatusIdle    readingStatus = "idle"
-	readingStatusPlaying readingStatus = "playing"
-	readingStatusPaused  readingStatus = "paused"
-)
-
-// ReadingState is an immutable snapshot of the reading module state suitable
-// for sending to a reconnecting client via the reading.state event.
-type ReadingState struct {
-	SectionID    string              `json:"sectionId"`
-	CurrentIndex int                 `json:"currentIndex"`
-	Lines        []readingLineConfig `json:"lines"`
-	BgmMediaID   string              `json:"bgmMediaId,omitempty"`
-	Status       string              `json:"status"`
-}
-
-type readingConfig struct {
-	AdvanceMode    string              `json:"AdvanceMode"`
-	DefaultVoiceID string              `json:"DefaultVoiceID"`
-	TotalLines     int                 `json:"TotalLines"`
-	Lines          []readingLineConfig `json:"Lines,omitempty"`
-	BGMId          string              `json:"BGMId,omitempty"`
-}
-
-type readingLineConfig struct {
-	Index        int    `json:"Index"`
-	Text         string `json:"Text,omitempty"`
-	VoiceID      string `json:"VoiceID,omitempty"`
-	AdvanceBy    string `json:"AdvanceBy,omitempty"`
-	Speaker      string `json:"Speaker,omitempty"`
-	VoiceMediaID string `json:"VoiceMediaID,omitempty"`
-}
-
-// validAdvanceBy reports whether s is a recognized advanceBy value:
-//   - "" (empty falls back to module-level advanceMode default; treated as "gm")
-//   - "gm"
-//   - "voice"
-//   - "role:<roleId>" with non-empty roleId
-func validAdvanceBy(s string) bool {
-	if s == "" || s == "gm" || s == "voice" {
-		return true
-	}
-	if strings.HasPrefix(s, "role:") && len(strings.TrimPrefix(s, "role:")) > 0 {
-		return true
-	}
-	return false
-}
-
-// resolveAdvanceBy returns the effective advanceBy for a given line index.
-// When the per-line value is empty it falls back to the module's advanceMode
-// (mapped to the per-line vocabulary).
-func (m *ReadingModule) resolveAdvanceBy(idx int) string {
-	if idx >= 0 && idx < len(m.lines) && m.lines[idx].AdvanceBy != "" {
-		return m.lines[idx].AdvanceBy
-	}
-	switch m.advanceMode {
-	case "auto":
-		return "voice"
-	case "player", "gm":
-		return "gm"
-	default:
-		return "gm"
-	}
-}
 
 // HandleVoiceEnded is the server-trusted advance entry point for lines whose
 // effective advanceBy is "voice". It bypasses the player/host permission
@@ -264,79 +175,6 @@ func (m *ReadingModule) HandleAdvance(ctx context.Context, playerID uuid.UUID, i
 		})
 	}
 	return nil
-}
-
-// NewReadingModule creates a new ReadingModule instance.
-func NewReadingModule() *ReadingModule {
-	return &ReadingModule{}
-}
-
-func (m *ReadingModule) Name() string { return "reading" }
-
-func (m *ReadingModule) Init(ctx context.Context, deps engine.ModuleDeps, config json.RawMessage) error {
-	m.mu.Lock()
-	defer m.mu.Unlock()
-
-	m.deps = deps
-
-	var cfg readingConfig
-	if config != nil && len(config) > 0 {
-		if err := json.Unmarshal(config, &cfg); err != nil {
-			return fmt.Errorf("reading: invalid config: %w", err)
-		}
-	}
-
-	// Apply defaults
-	if cfg.AdvanceMode == "" {
-		cfg.AdvanceMode = "gm"
-	}
-
-	m.advanceMode = cfg.AdvanceMode
-	m.defaultVoiceID = cfg.DefaultVoiceID
-	m.bgmId = cfg.BGMId
-	m.lines = cfg.Lines
-	m.totalLines = cfg.TotalLines
-	if m.totalLines == 0 && len(m.lines) > 0 {
-		m.totalLines = len(m.lines)
-	}
-	m.currentLineIndex = 0
-	m.isActive = true
-	m.status = readingStatusPlaying
-	m.pausedReason = ""
-
-	// If section BGM configured, emit override on activation
-	if m.bgmId != "" {
-		deps.EventBus.Publish(engine.Event{
-			Type:    "audio.set_bgm",
-			Payload: map[string]any{"mediaId": m.bgmId, "fadeMs": 1000},
-		})
-	}
-
-	// Emit reading.started so the WS bridge can push a reading:started
-	// envelope to all connected session clients. Lines are emitted in the
-	// raw storage (PascalCase) shape; the bridge converts to camelCase wire
-	// format before broadcast. A defensive copy is taken so downstream
-	// subscribers cannot mutate module-owned state.
-	linesCopy := make([]readingLineConfig, len(m.lines))
-	copy(linesCopy, m.lines)
-	deps.EventBus.Publish(engine.Event{
-		Type: "reading.started",
-		Payload: map[string]any{
-			"lines":      linesCopy,
-			"bgmMediaId": m.bgmId,
-			"totalLines": m.totalLines,
-		},
-	})
-
-	return nil
-}
-
-// resolveVoiceID returns the voiceID for a given line index, falling back to default.
-func (m *ReadingModule) resolveVoiceID(idx int) string {
-	if idx >= 0 && idx < len(m.lines) && m.lines[idx].VoiceID != "" {
-		return m.lines[idx].VoiceID
-	}
-	return m.defaultVoiceID
 }
 
 // HandleMessage is the legacy engine.Module entry point for reading messages.
@@ -511,142 +349,4 @@ func (m *ReadingModule) HandlePlayerRejoined(ctx context.Context, hostRejoining 
 			"advanceBy": advanceBy,
 		},
 	})
-}
-
-// GetState returns an immutable snapshot of the current reading state for
-// reconnecting clients. The Lines slice is copied so that callers cannot
-// mutate internal module state.
-func (m *ReadingModule) GetState() ReadingState {
-	m.mu.RLock()
-	defer m.mu.RUnlock()
-
-	linesCopy := make([]readingLineConfig, len(m.lines))
-	copy(linesCopy, m.lines)
-
-	status := string(m.status)
-	if status == "" {
-		if m.isActive {
-			status = string(readingStatusPlaying)
-		} else {
-			status = string(readingStatusIdle)
-		}
-	}
-
-	return ReadingState{
-		// SectionID is reserved for future use; the reading module currently
-		// does not track its owning section identifier — the WS handler can
-		// inject it when broadcasting if needed.
-		SectionID:    "",
-		CurrentIndex: m.currentLineIndex,
-		Lines:        linesCopy,
-		BgmMediaID:   m.bgmId,
-		Status:       status,
-	}
-}
-
-// ReadingStateWire is a JSON-encoding-friendly snapshot of the reading state
-// suitable for sending across the WebSocket without leaking unexported types.
-// Lines is pre-marshalled to json.RawMessage so that ws.ReadingStateSnapshot
-// (which uses RawMessage) can ferry it through unchanged.
-type ReadingStateWire struct {
-	SectionID    string          `json:"sectionId"`
-	CurrentIndex int             `json:"currentIndex"`
-	Lines        json.RawMessage `json:"lines"`
-	BgmMediaID   string          `json:"bgmMediaId,omitempty"`
-	Status       string          `json:"status"`
-}
-
-// GetReadingStateWire returns a wire-friendly snapshot of the current
-// reading state. Lines are pre-marshalled so callers don't need to know
-// about the unexported readingLineConfig type. The SectionID field is left
-// empty here; the WS bridge injects it from the per-session registry.
-func (m *ReadingModule) GetReadingStateWire() ReadingStateWire {
-	state := m.GetState()
-	linesJSON, err := json.Marshal(state.Lines)
-	if err != nil {
-		// json.Marshal of a known struct slice should never fail; fall back
-		// to an empty array so the wire envelope stays valid JSON.
-		linesJSON = json.RawMessage(`[]`)
-	}
-	return ReadingStateWire{
-		SectionID:    state.SectionID,
-		CurrentIndex: state.CurrentIndex,
-		Lines:        linesJSON,
-		BgmMediaID:   state.BgmMediaID,
-		Status:       state.Status,
-	}
-}
-
-func (m *ReadingModule) BuildState() (json.RawMessage, error) {
-	m.mu.RLock()
-	defer m.mu.RUnlock()
-
-	state := map[string]any{
-		"currentLine": m.currentLineIndex,
-		"totalLines":  m.totalLines,
-		"isActive":    m.isActive,
-		"advanceMode": m.advanceMode,
-	}
-	return json.Marshal(state)
-}
-
-func (m *ReadingModule) Cleanup(ctx context.Context) error {
-	m.mu.Lock()
-	defer m.mu.Unlock()
-
-	m.isActive = false
-	return nil
-}
-
-// Schema returns the JSON Schema for ReadingModule settings.
-func (m *ReadingModule) Schema() json.RawMessage {
-	return json.RawMessage(`{
-		"type": "object",
-		"properties": {
-			"AdvanceMode": {"type": "string", "enum": ["gm", "auto", "player"], "default": "gm"},
-			"DefaultVoiceID": {"type": "string"},
-			"TotalLines": {"type": "integer", "minimum": 0},
-			"BGMId": {"type": "string"},
-			"Lines": {
-				"type": "array",
-				"items": {
-					"type": "object",
-					"properties": {
-						"VoiceID": {"type": "string"}
-					}
-				}
-			}
-		}
-	}`)
-}
-
-// --- PhaseHookModule ---
-
-func (m *ReadingModule) OnPhaseEnter(_ context.Context, _ engine.Phase) error {
-	return nil
-}
-
-func (m *ReadingModule) OnPhaseExit(_ context.Context, _ engine.Phase) error {
-	return nil
-}
-
-// BuildStateFor returns the same state as BuildState for now.
-// PR-2a (F-sec-2 gate): satisfies engine.PlayerAwareModule interface.
-// PR-2b will tailor the line window to the viewer — for role-gated advance
-// the current line's prompts should be restricted to the owning role, and
-// future lines should be redacted until revealed.
-func (m *ReadingModule) BuildStateFor(_ uuid.UUID) (json.RawMessage, error) {
-	return m.BuildState()
-}
-
-// Compile-time interface checks.
-var (
-	_ engine.Module            = (*ReadingModule)(nil)
-	_ engine.ConfigSchema      = (*ReadingModule)(nil)
-	_ engine.PhaseHookModule   = (*ReadingModule)(nil)
-	_ engine.PlayerAwareModule = (*ReadingModule)(nil)
-)
-
-func init() {
-	engine.Register("reading", func() engine.Module { return NewReadingModule() })
 }
